fix: update FilmMapper

+++ src/main/java/ru/yandex/practicum/filmorate/model/film/Genre.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Genre {
-    private byte id;
+    private short id;
     private String name;
 }
